// frontend/src/types/entertainer.ts
export interface Entertainer {
    entertainerID: number;
    entStageName: string;
    entssn: string;
    entStreetAddress: string;
    entCity: string;
    entState: string;
    entZipCode: string;
    entPhoneNumber: string;
    entWebPage: string;
    entEMailAddress: string;
    dateEntered: string;
}